# swagger/doc.yml
swagger: "2.0"
info:
  description: "A GinBro RESTful APIs"
  version: "1.0.0"
  title: "GinBro RESTful APIs Application"
host: "ynov-api.ew.r.appspot.com"
basePath: "/api/v1"
schemes:
- "http"
paths:
...
  /real-estate:
    get:
      tags:
      - "real-estate"
      summary: "get all real-estate by pagination"
      description: ""
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "where"
        in: "query"
        description: "column:value will use sql LIKE for search eg:id:67 will where id >67 eg2: name:eric => where name LIKE '%eric%'"
        required: false
        type: "array"
        items:
          type: "string"
      - name: "fields"
        in: "query"
        description: "{$tableColumn},{$tableColumn}... "
        required: false
        type: "string"
      - name: "order"
        in: "query"
        description: "eg: id desc, name desc"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "sql offset eg: 10"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        default: "2"
        description: "limit returning object count"
        required: false
        type: "integer"

      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/RealEstatePagination"
    post:
      tags:
      - "real-estate"
      summary: "create real-estate"
      description: "create real-estate"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create real-estate"
        required: true
        schema:
          $ref: "#/definitions/RealEstate"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"

    patch:
      tags:
      - "real-estate"
      summary: "update real-estate"
      description: "update real-estate"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create real-estate"
        required: true
        schema:
          $ref: "#/definitions/RealEstate"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
  /real-estate/{Id}:
    get:
      tags:
      - "real-estate"
      summary: "get a real-estate by ID"
      description: "get a real-estate by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of real-estate to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/RealEstate"
    delete:
      tags:
      - "real-estate"
      summary: "Destroy a real-estate by ID"
      description: "delete a real-estate by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of real-estate to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
  /real-estate-availability:
    get:
      tags:
      - "real-estate-availability"
      summary: "get all real-estate-availability by pagination"
      description: ""
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "where"
        in: "query"
        description: "column:value will use sql LIKE for search eg:id:67 will where id >67 eg2: name:eric => where name LIKE '%eric%'"
        required: false
        type: "array"
        items:
          type: "string"
      - name: "fields"
        in: "query"
        description: "{$tableColumn},{$tableColumn}... "
        required: false
        type: "string"
      - name: "order"
        in: "query"
        description: "eg: id desc, name desc"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "sql offset eg: 10"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        default: "2"
        description: "limit returning object count"
        required: false
        type: "integer"

      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/RealEstateAvailabilityPagination"
    post:
      tags:
      - "real-estate-availability"
      summary: "create real-estate-availability"
      description: "create real-estate-availability"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create real-estate-availability"
        required: true
        schema:
          $ref: "#/definitions/RealEstateAvailability"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"

    patch:
      tags:
      - "real-estate-availability"
      summary: "update real-estate-availability"
      description: "update real-estate-availability"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create real-estate-availability"
        required: true
        schema:
          $ref: "#/definitions/RealEstateAvailability"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
  /real-estate-availability/{Id}:
    get:
      tags:
      - "real-estate-availability"
      summary: "get a real-estate-availability by ID"
      description: "get a real-estate-availability by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of real-estate-availability to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/RealEstateAvailability"
    delete:
      tags:
      - "real-estate-availability"
      summary: "Destroy a real-estate-availability by ID"
      description: "delete a real-estate-availability by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of real-estate-availability to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
  /real-estate-feedback:
    get:
      tags:
      - "real-estate-feedback"
      summary: "get all real-estate-feedback by pagination"
      description: ""
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "where"
        in: "query"
        description: "column:value will use sql LIKE for search eg:id:67 will where id >67 eg2: name:eric => where name LIKE '%eric%'"
        required: false
        type: "array"
        items:
          type: "string"
      - name: "fields"
        in: "query"
        description: "{$tableColumn},{$tableColumn}... "
        required: false
        type: "string"
      - name: "order"
        in: "query"
        description: "eg: id desc, name desc"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "sql offset eg: 10"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        default: "2"
        description: "limit returning object count"
        required: false
        type: "integer"

      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/RealEstateFeedbackPagination"
    post:
      tags:
      - "real-estate-feedback"
      summary: "create real-estate-feedback"
      description: "create real-estate-feedback"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create real-estate-feedback"
        required: true
        schema:
          $ref: "#/definitions/RealEstateFeedback"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"

    patch:
      tags:
      - "real-estate-feedback"
      summary: "update real-estate-feedback"
      description: "update real-estate-feedback"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create real-estate-feedback"
        required: true
        schema:
          $ref: "#/definitions/RealEstateFeedback"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
  /real-estate-feedback/{Id}:
    get:
      tags:
      - "real-estate-feedback"
      summary: "get a real-estate-feedback by ID"
      description: "get a real-estate-feedback by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of real-estate-feedback to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/RealEstateFeedback"
    delete:
      tags:
      - "real-estate-feedback"
      summary: "Destroy a real-estate-feedback by ID"
      description: "delete a real-estate-feedback by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of real-estate-feedback to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
  /real-estate-image:
    get:
      tags:
      - "real-estate-image"
      summary: "get all real-estate-image by pagination"
      description: ""
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "where"
        in: "query"
        description: "column:value will use sql LIKE for search eg:id:67 will where id >67 eg2: name:eric => where name LIKE '%eric%'"
        required: false
        type: "array"
        items:
          type: "string"
      - name: "fields"
        in: "query"
        description: "{$tableColumn},{$tableColumn}... "
        required: false
        type: "string"
      - name: "order"
        in: "query"
        description: "eg: id desc, name desc"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "sql offset eg: 10"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        default: "2"
        description: "limit returning object count"
        required: false
        type: "integer"

      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/RealEstateImagePagination"
    post:
      tags:
      - "real-estate-image"
      summary: "create real-estate-image"
      description: "create real-estate-image"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create real-estate-image"
        required: true
        schema:
          $ref: "#/definitions/RealEstateImage"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"

    patch:
      tags:
      - "real-estate-image"
      summary: "update real-estate-image"
      description: "update real-estate-image"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create real-estate-image"
        required: true
        schema:
          $ref: "#/definitions/RealEstateImage"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
  /real-estate-image/{Id}:
    get:
      tags:
      - "real-estate-image"
      summary: "get a real-estate-image by ID"
      description: "get a real-estate-image by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of real-estate-image to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/RealEstateImage"
    delete:
      tags:
      - "real-estate-image"
      summary: "Destroy a real-estate-image by ID"
      description: "delete a real-estate-image by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of real-estate-image to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
...
definitions:
...
  RealEstate:
    type: "object"
    properties:
    
      id:
        type: "integer"
        description: "Id du bien"
        format: "int64"
      
      id_user:
        type: "integer"
        description: "Id du proprietaire"
        format: "int64"
      
      accroche:
        type: "string"
        description: "Phrase d'accroche (max 100 caractères)"
        format: "string"
      
      type:
        type: "string"
        description: "Type de bien ('apartment','house','vacant_lot','parking','loft','castle','building')"
        format: "string"
      
      nb_rooms:
        type: "integer"
        description: "Nombre de pièce"
        format: "int64"
      
      nb_bedroom:
        type: "integer"
        description: "Nombre de chambre"
        format: "int64"
      
      description:
        type: "string"
        description: "Description"
        format: "string"
      
      size:
        type: "integer"
        description: "Superficie (en entier)"
        format: "int64"
      
      price:
        type: "integer"
        description: "Prix du bien"
        format: "int64"
      
      address:
        type: "string"
        description: "Adresse"
        format: "string"
      
      zip_code:
        type: "string"
        description: "Code postal"
        format: "string"
      
      city:
        type: "string"
        description: "Ville"
        format: "string"
      
      latitude:
        type: "string"
        description: "Latitude"
        format: "string"
      
      longitude:
        type: "string"
        description: "Longitude"
        format: "string"
      
      energy_class:
        type: "string"
        description: "Classe d'énergie"
        format: "string"
      
      ges_class:
        type: "string"
        description: "Classe GES"
        format: "string"
      
      has_garden:
        type: "integer"
        description: "A un jardin"
        format: "int64"
      
      has_exposed_stone:
        type: "integer"
        description: "A des pierres apparentes"
        format: "int64"
      
      has_ciment_tiles:
        type: "integer"
        description: "A des carreaux de ciments"
        format: "int64"
      
      has_parquet_floor:
        type: "integer"
        description: "A du parquet au sol"
        format: "int64"
  RealEstatePagination:
    type: "object"
    properties:
      code:
        type: "integer"
        description: "json repose code"
        format: "int32"
      total:
        type: "integer"
        description: "total numbers"
        format: "int32"
      offset:
        type: "integer"
        description: "offset"
        format: "int32"
      limit:
        type: "integer"
        description: "limit"
        format: "int32"
      list:
        type: "array"
        items:
          $ref: "#/definitions/RealEstate"
  RealEstateAvailability:
    type: "object"
    properties:
    
      id:
        type: "integer"
        description: "Id de la disponibilité"
        format: "int64"
      
      price:
        type: "number"
        description: "Prix de la visite"
        format: "float"
      
      start_date:
        type: "string"
        description: "Date de début"
        format: "date-time"
      
      end_date:
        type: "string"
        description: "Date de fin"
        format: "date-time"
      
      start_time:
        type: "string"
        description: "De "
        format: "date-time"
      
      end_time:
        type: "string"
        description: "à"
        format: "date-time"
  RealEstateAvailabilityPagination:
    type: "object"
    properties:
      code:
        type: "integer"
        description: "json repose code"
        format: "int32"
      total:
        type: "integer"
        description: "total numbers"
        format: "int32"
      offset:
        type: "integer"
        description: "offset"
        format: "int32"
      limit:
        type: "integer"
        description: "limit"
        format: "int32"
      list:
        type: "array"
        items:
          $ref: "#/definitions/RealEstateAvailability"
  RealEstateFeedback:
    type: "object"
    properties:
    
      id:
        type: "integer"
        description: "Id du feedback"
        format: "int64"
      
      id_user:
        type: "integer"
        description: "Id de l'utilisateur"
        format: "int64"
      
      id_real_estate:
        type: "integer"
        description: "Id du bien"
        format: "int64"
      
      feedback:
        type: "string"
        description: "Feedback"
        format: "string"
      
      date:
        type: "string"
        description: "Date (YYYY-mm-dd hh:mm:ss)"
        format: "date-time"
  RealEstateFeedbackPagination:
    type: "object"
    properties:
      code:
        type: "integer"
        description: "json repose code"
        format: "int32"
      total:
        type: "integer"
        description: "total numbers"
        format: "int32"
      offset:
        type: "integer"
        description: "offset"
        format: "int32"
      limit:
        type: "integer"
        description: "limit"
        format: "int32"
      list:
        type: "array"
        items:
          $ref: "#/definitions/RealEstateFeedback"
  RealEstateImage:
    type: "object"
    properties:
    
      id:
        type: "integer"
        description: "Id de l'image"
        format: "int64"
      
      id_real_estate:
        type: "integer"
        description: "Id du bien"
        format: "int64"
      
      url:
        type: "string"
        description: "Url de l'image"
        format: "string"
  RealEstateImagePagination:
    type: "object"
    properties:
      code:
        type: "integer"
        description: "json repose code"
        format: "int32"
      total:
        type: "integer"
        description: "total numbers"
        format: "int32"
      offset:
        type: "integer"
        description: "offset"
        format: "int32"
      limit:
        type: "integer"
        description: "limit"
        format: "int32"
      list:
        type: "array"
        items:
          $ref: "#/definitions/RealEstateImage"
...
  ApiResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      msg:
        type: "string"
...